+++ build-utils/cmake/compiler_settings.cmake
@@ -60,6 +60,15 @@ set(cxx_release_flags
   CMAKE_C_FLAGS_RELEASE)
 
 #--------------------------------------
+# Cross compiler compilation flags
+
+# Simd math force ref
+if(ozz_build_simd_ref)
+  message(STATUS "Forces SimdMath reference implementation")
+  set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS OZZ_FORCE_SIMD_REF)
+endif()
+
+#--------------------------------------
 # Modify default MSVC compilation flags
 if(MSVC)
 
@@ -69,36 +78,20 @@ if(MSVC)
   # Disables crt secure warnings
   set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS _CRT_SECURE_NO_WARNINGS)
 
-  # SSE detection
-  if(ozz_build_sse2 OR CMAKE_CL_64) # x64 implicitly supports SSE2.
-    message("OZZ_HAS_SSE2 is enabled")
-    set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS OZZ_HAS_SSE2=1)
-  endif()
-
   # Adds support for SSE instructions
   string(REGEX REPLACE " /arch:SSE[0-9]?" "" CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
   string(REGEX REPLACE " /arch:SSE[0-9]?" "" CMAKE_C_FLAGS "${CMAKE_C_FLAGS}")
-  if(ozz_build_sse2 AND NOT CMAKE_CL_64) # x64 implicitly supports SSE2.
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /arch:SSE2")
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} /arch:SSE2")
+  if(NOT ozz_build_simd_ref AND NOT CMAKE_CL_64)  # /arch:SSE2 isn't valid for x64 builds.
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "/arch:SSE2")
   endif()
 
   # Removes any exception mode
   string(REGEX REPLACE " /EH.*" "" CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
   string(REGEX REPLACE " /EH.*" "" CMAKE_C_FLAGS "${CMAKE_C_FLAGS}")
-
-  # Disables STL exceptions also
-  if(NOT ${CMAKE_CXX_FLAGS} MATCHES "/D _HAS_EXCEPTIONS=0")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /D _HAS_EXCEPTIONS=0")
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} /D _HAS_EXCEPTIONS=0")
-  endif()
+  set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS _HAS_EXCEPTIONS=0)
   
-
   # Adds support for multiple processes builds
-  if(NOT ${CMAKE_CXX_FLAGS} MATCHES "/MP")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /MP")
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} /MP")
-  endif()
+  set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "/MP")
 
   #---------------
   # For all builds
@@ -114,13 +107,8 @@ if(MSVC)
   endforeach()
   
   #----------------------
-  # For debug builds only
-  foreach(flag ${cxx_debug_flags})
-    # Enables smaller type checks
-    if(NOT ${flag} MATCHES "/RTCc")
-      set(${flag} "${${flag}} /RTCc")
-    endif()
-  endforeach()
+  # Enables smaller type checks for debug builds only
+  set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS $<$<CONFIG:DEBUG>:/RTCc>)
 
 #--------------------------------------
 # else consider the compiler as GCC compatible
@@ -132,31 +120,19 @@ else()
 
   # Enable c++11
   if(ozz_build_cpp11)
-    if(NOT CMAKE_CXX_FLAGS MATCHES "-std=c\\+\\+11")
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
-    endif()
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "$<$<STREQUAL:$<TARGET_PROPERTY:LINKER_LANGUAGE>,CXX>:-std=c++11>")
   endif()
 
   # Set the warning level to Wall
-  if(NOT CMAKE_CXX_FLAGS MATCHES "-Wall")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall")
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall")
-  endif()
+  set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "-Wall")
 
   # Automatically selects native architecture optimizations (sse...)
-  #if(NOT CMAKE_CXX_FLAGS MATCHES "-march")
-  #  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native")
-  #  set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -march=native")
-  #endif()
+  #set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "-march=native")
 
   #----------------------
   # Enables debug glibcxx if NDebug isn't defined, not supported by APPLE
   if(NOT APPLE)
-    foreach(flag ${cxx_all_but_default_flags})
-      if(NOT ${flag} MATCHES "NDEBUG" AND NOT ${flag} MATCHES "-D_GLIBCXX_DEBUG")
-        set(${flag} "${${flag}} -D_GLIBCXX_DEBUG")
-      endif()
-    endforeach()
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS $<$<CONFIG:Debug>:_GLIBCXX_DEBUG>)
   endif()
 
   #----------------------
@@ -166,17 +142,11 @@ else()
 
     #----------------------
     # Disable emscripten absolute-paths warning
-    if(NOT CMAKE_CXX_FLAGS MATCHES "-Wno-warn-absolute-paths")
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-warn-absolute-paths")
-      set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wno-warn-absolute-paths")
-    endif()
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "-Wno-warn-absolute-paths")
 
     #----------------------
     # Sets emscripten output
-    if(NOT CMAKE_CXX_FLAGS MATCHES "-s DISABLE_GL_EMULATION=1")
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -s DISABLE_GL_EMULATION=1")
-      set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -s DISABLE_GL_EMULATION=1")
-    endif()
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "-s DISABLE_GL_EMULATION=1")
   endif()
 
   #----------------------
@@ -194,32 +164,37 @@ else()
     endif()
 
     # Compiler options
-    foreach(flag ${cxx_debug_flags})
-      if(NOT ${flag} MATCHES "-fprofile-arcs -ftest-coverage")
-        set(${flag} "${${flag}} -fprofile-arcs -ftest-coverage")
-      endif()
-    endforeach()
+    set_property(DIRECTORY APPEND PROPERTY COMPILE_OPTIONS "-fprofile-arcs -ftest-coverage")
   endif()
 
 endif()
 
 #---------------------
 # Prints all the flags
-message("---------------------------------------------------------")
-message("Default build type is: ${CMAKE_BUILD_TYPE}")
-message("The following compilation flags will be used:")
+message(STATUS "---------------------------------------------------------")
+message(STATUS "Default build type is: ${CMAKE_BUILD_TYPE}")
+message(STATUS "The following compilation flags will be used:")
 foreach(flag ${cxx_all_flags})
   message(${flag} " ${${flag}}")
 endforeach()
-message("---------------------------------------------------------")
 
-#-----------------------------------------------
-# Updates cmake cache, will be visible in the UI
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}" CACHE STRING "Flags used by the compiler for all buids" FORCE)
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG}" CACHE STRING "Flags used by the compiler for debug buids" FORCE)
-set(CMAKE_CXX_FLAGS_MINSIZEREL "${CMAKE_CXX_FLAGS_MINSIZEREL}" CACHE STRING "Flags used by the compiler for MinSizeRel buids" FORCE)
-set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${CMAKE_CXX_FLAGS_RELWITHDEBINFO}" CACHE STRING "Flags used by the compiler for RelWithDebugInfo buids" FORCE)
-set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE}" CACHE STRING "Flags used by the compiler for release buids" FORCE)
+message(STATUS "---------------------------------------------------------")
+
+get_directory_property(DirectoryCompileOptions DIRECTORY ${CMAKE_SOURCE_DIR} COMPILE_OPTIONS)
+message(STATUS "Directory Compile Options:")
+foreach(opt ${DirectoryCompileOptions})
+  message(STATUS ${opt})
+endforeach()
+
+message(STATUS "---------------------------------------------------------")
+
+get_directory_property(DirectoryCompileDefinitions DIRECTORY ${CMAKE_SOURCE_DIR} COMPILE_DEFINITIONS)
+message(STATUS "Directory Compile Definitions:")
+foreach(def ${DirectoryCompileDefinitions})
+  message(STATUS ${def})
+endforeach()
+
+message(STATUS "---------------------------------------------------------")
 
 #----------------------------------------------
 # Modifies output directory for all executables